refactor: Rename studioDefines.js to engineDefines.js
