+++ src/gizmos/GizmoManager.js
@@ -1,7 +1,7 @@
 import { Entity } from "../core/Entity.js";
 import { Material } from "../rendering/Material.js";
 import { VertexState } from "../rendering/VertexState.js";
-import { ENGINE_ASSETS_LIVE_UPDATES_SUPPORT } from "../studioDefines.js";
+import { ENGINE_ASSETS_LIVE_UPDATES_SUPPORT } from "../engineDefines.js";
 import { RotateAxisGizmoDraggable } from "./draggables/RotateAxisGizmoDraggable.js";
 import { ScaleGizmoDraggable } from "./draggables/ScaleGizmoDraggable.js";
 import { TranslateAxisGizmoDraggable } from "./draggables/TranslateAxisGizmoDraggable.js";
